Misc updates

+++ c/tests/test_stats.c
@@ -1065,7 +1065,7 @@ test_paper_ex_diversity(void)
     ret = tsk_treeseq_diversity(&ts, 1, &sample_set_sizes, samples, 0, NULL,
             &pi1, TSK_STAT_SITE);
     CU_ASSERT_EQUAL_FATAL(ret, 0);
-    CU_ASSERT(isnan(pi1));
+    CU_ASSERT(isnan((float) pi1));
 
     tsk_treeseq_free(&ts);
 }
@@ -1216,7 +1216,7 @@ test_paper_ex_Y1(void)
     sample_set_sizes = 1;
     ret = tsk_treeseq_Y1(&ts, 1, &sample_set_sizes, samples, 0, NULL, &result, 0);
     CU_ASSERT_EQUAL_FATAL(ret, 0);
-    CU_ASSERT(isnan(result));
+    CU_ASSERT(isnan((float) result));
 
     tsk_treeseq_free(&ts);
 }
@@ -1255,7 +1255,7 @@ test_paper_ex_divergence(void)
     ret = tsk_treeseq_divergence(&ts, 2, sample_set_sizes, samples,
             1, set_indexes, 0, NULL, &result, TSK_STAT_SITE);
     CU_ASSERT_EQUAL_FATAL(ret, 0);
-    CU_ASSERT(isnan(result));
+    CU_ASSERT(isnan((float) result));
 
     tsk_treeseq_free(&ts);
 }
@@ -1293,7 +1293,7 @@ test_paper_ex_Y2(void)
     ret = tsk_treeseq_Y2(&ts, 2, sample_set_sizes, samples,
             1, set_indexes, 0, NULL, &result, TSK_STAT_SITE);
     CU_ASSERT_EQUAL_FATAL(ret, 0);
-    CU_ASSERT(isnan(result));
+    CU_ASSERT(isnan((float) result));
 
     tsk_treeseq_free(&ts);
 }
@@ -1331,7 +1331,7 @@ test_paper_ex_f2(void)
     ret = tsk_treeseq_f2(&ts, 2, sample_set_sizes, samples,
             1, set_indexes, 0, NULL, &result, TSK_STAT_SITE);
     CU_ASSERT_EQUAL_FATAL(ret, 0);
-    CU_ASSERT(isnan(result));
+    CU_ASSERT(isnan((float) result));
 
     /* sample_set_size of 1 leads to NaN */
     sample_set_sizes[0] = 2;
@@ -1339,7 +1339,7 @@ test_paper_ex_f2(void)
     ret = tsk_treeseq_f2(&ts, 2, sample_set_sizes, samples,
             1, set_indexes, 0, NULL, &result, TSK_STAT_SITE);
     CU_ASSERT_EQUAL_FATAL(ret, 0);
-    CU_ASSERT(isnan(result));
+    CU_ASSERT(isnan((float) result));
 
     tsk_treeseq_free(&ts);
 }
@@ -1408,7 +1408,7 @@ test_paper_ex_f3(void)
     ret = tsk_treeseq_f3(&ts, 3, sample_set_sizes, samples,
             1, set_indexes, 0, NULL, &result, TSK_STAT_SITE);
     CU_ASSERT_EQUAL_FATAL(ret, 0);
-    CU_ASSERT(isnan(result));
+    CU_ASSERT(isnan((float) result));
 
     tsk_treeseq_free(&ts);
 }
